Introduce hopper to the world
This library is motivated by the need to collect telemetry data
from bursty systems and not lose any bit of it _and_ also not
consume unbounded amounts of memory on the collecting system. To
that end, hopper acts as an mpsc but pages items out to disk in
burst situations.

hopper is intended to be used in any situation where an unbounded
mpsc from the stdlib would be used but you must be able to
guarantee fixed memory consumption.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>