AFL suggested corrections, plus runner in resources
This commit has a small number of changes to the code from the last
commit. We now handle the case where the max_bytes is too small to
hold the size of the type T of the sender / receiver pair and default
to the minimum size needed.

I've also added the AFL runner under resources/. It will not be run
automatically.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>