Use write bounds to denote non-continguous writes (#5)
* Use write bounds to denote non-continguous writes

Previously hopper suffered fragmentation problems in its array of
memories. In particular, it was possible for a receiver to lose
writes and re-detect them at a later time in an out of order fashion.
The basic access pattern of hopper has always assumed a continually
growing "index" in memory, which was disk. With the introduction of
the _actual_ in-memory buffers it was possible for either the sender
or the receiver to backtrack and then skip a chunk of memory to
avoid reading in something from a later write.

The solution is to store the bounds of writes. A receiver keeps
a rolling ID of its reads which the senders use to note when a
new bound is needed, owing, potentially, to the current bound
having been invalidated. So long as the sender captured receiver
id remains the same the current write boundaries grow.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* AFL suggested corrections, plus runner in resources

This commit has a small number of changes to the code from the last
commit. We now handle the case where the max_bytes is too small to
hold the size of the type T of the sender / receiver pair and default
to the minimum size needed.

I've also added the AFL runner under resources/. It will not be run
automatically.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>