Bump to 0.1.2
This release corrects the out-of-order delivery issue that was
introduced in 0.1.1. Hopper now keeps an explicit list of write
indecies for the receiver to read from. This is potentially an
unbounded list--as we cannot toss any writes--but the overhead
is one two-tuple of usize per index bound and there should be
progress made between the receiver and senders to cause that
list to be practically bounded.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>