Bump bincode dep to -alpha6, add fuzz tests (#12)
The API for bincode changed between alpha2 and alpha6, causing
build failures for any system that depends on hopper and does a
'cargo update'. Hopper now uses the latest bincode and we verify
that these changes don't cause heartache by porting the old AFL
tests to cargo-fuzz. Instructions for fiddling with the docker
container to run these tests are in fuzz/README

Signed-off-by: Brian L. Troutwine <blt@postmates.com>