Upgrade to Bincode 0.8.0, Serde 1.0 (#13)
* Run cargo fmt

This commit only fiddles with the layout of the source code. There's
no bigger thing that happens.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Update to bincode 0.8.0, serde 1.0

This commit bumps the bincode dep to 0.8.0 which brings the
dependency on serde up to 1.0. As a result the Deserialize
trait we used to use now takes a lifetime. I _think_ we want
DeserializeOwned on account of we want to throw the intermediate
buffer away, per the justification for DeserializeOwned here:

    https://serde.rs/lifetimes.html

It is a pain, though, that we're continually constructing that
little vector, copying some parts out -- depending on the T --
and tossing the whole thing. But! I guess we can't do much
better since we'd need to keep whatever we have resident in place
somewhere in memory to satisfy the liftime of the T.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Clippy fixes

This commit is due to clippy lints. There's no effect change but
things are a little cleaner, especially in afl_crashes.rs.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>