Begin work to remove unbounded write_bounds
We've known for some time that write_bounds could become unbounded.
It turns out, in practice, it is. This makes the comment of "yikes,
unbounded" even more amusing. The accumulation is very slow, 128 bytes
every so often but it's enough.

Just goes to show, you should never allow unbounded collections.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>