Fully remove unbounded VecDeque
We still need write_bounds to control when the senders set the
read index for the receiver they are bound to but we have no need
to buffer them into a VecDeque. In fact, owing to the pushing into
the front of the buffer and popping only from the back a single
Option is sufficient.

This removes the unbounded member of the fslock structure.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>