Removed unbounded member of FsLock (#14)
* Begin work to remove unbounded write_bounds

We've known for some time that write_bounds could become unbounded.
It turns out, in practice, it is. This makes the comment of "yikes,
unbounded" even more amusing. The accumulation is very slow, 128 bytes
every so often but it's enough.

Just goes to show, you should never allow unbounded collections.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Fully remove unbounded VecDeque

We still need write_bounds to control when the senders set the
read index for the receiver they are bound to but we have no need
to buffer them into a VecDeque. In fact, owing to the pushing into
the front of the buffer and popping only from the back a single
Option is sufficient.

This removes the unbounded member of the fslock structure.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>