Further trimmed down the FsSync
I have done a fair bit of trimming of FsSync. We no longer have
_so_ much repetitive fiddling going on in that shared structure
and the sender / receiver are simpler for it.

I _hope_ to get the FsSync totally destroyed. The only things that
remain are a count of _disk_ writes and the VecDeque. I don't have
a plan per se but maybe one will come to me.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>