Remove afl tests
This is causing -- I think -- travis to spin out. It does _not_
fail on my local machine, probably because travis disks are very
slow.

As a part of this work I'm going to add proper AFL testing back
into the project. They won't run in travis but they'll be there.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>