This locks, but doesn't segfault
Okey dokey, so this is some kind of progress. What's happening here
is it's possible for the enq to lose the thread if it rumbles oddly
with deq.

The thing is, this won't square with my need to attempt an enq,
get bounced and then deq_n. I need to invest in contiguous storage
which I believe will be better for me anyhow as I can press put
the enq/deq index into the enq_lock / deq_lock mutex, this skip out
on the condvar entirely.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>