Passing tests on deque
We still yet need to do an AFL run and longer-duration QC runs as
well but this is progress. The primary change is we no longer point
to contiguous options but to const pointers, being sure to synchronize
on size as AoMP suggests for ARM and friends' sake.

Special thanks to https://users.rust-lang.org/t/why-do-my-distinct-values-resolve-to-the-same-pointer/15517/3
for help figuring out my stack allocation issue.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>