Worked myself into a corner, but closer to done
I've currently got a double mut issue with the way deque works
but I'm closer to where I need to be. Namely, the sender / receiver
_should_ have fine grained control over their locks and the sender
loop is closer to functional. Now I just have to figure out how
to take a lock and then meddle with it. I think the answer is
pushing InnerQueue back up into Queue.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>