Peel back some of the mutable borrow errors
This commit manages to work back some of the mutable borrow
errors that plauged the last commit. The main difficulty now
is located entirely in Sender, which, uh, is progress. Somehow
I need to fiddle with write_to_disk so that we can do rotations
without having to write to self, or something.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>