Only delete queue files at channel creation
Prior to this commit the Receiver was responsible for nuking
unused queue files before starting up. Consider what happens if
the receiver is very, very late to the party. It'll toast the
files it needs to read, then fall over when the files it needs
to read are gone.

Oops!

Signed-off-by: Brian L. Troutwine <blt@postmates.com>