Demonstrate locking failure
It turns out there's a bug still yet in the code. It's possible to
cause the Receiver to lock up. This happens when the queue is full
and we pull two items, even though the atomic counter had previously
been incremented to show the presence of one of those items. The
Receiver and Sender then double-pop -- in effect -- the item.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>