Vauge progress
Well, there's progress here of a kind. pop_back is now gone, we
only have two operations on the deque and once something's in there
it's in there until its permanently gone. Good. We also block
the sender on count size rather than in-memory representation. Also
good.

Issue is, tests still fail and the need for a flush command to drive
the last disk writes out is a tedious API. Only needed on shutdown, flush,
so that's a positive for actual Hopper use.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>