Include two explicit tests
One issue we had with the previous now-we-are-done commit was that
none of our tests looped. For any given input it's possible that
a concurrency bug will force failure in 1/10000 runs, say. Which,
boo-urns, right?

This commit does not include any 10k loops but does have a 2.5k loop.
The cautious developer is encouraged to bulk this up during development.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>