Clean up for review
Okey dokey! I'm pretty sure at this point we're back on track. I've
passed 1M QC tests plus the new loopers and the deque is about as
bare-bones as I can conceive making it at this point, which is a good
sign I do believe. Mostly this commit is fiddling with documentation.
I did adjust the benchmark to have the same ith values so we could
compare apples to apples.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>