Add a cap on total queue files
This commit adds a `max_disk_files` option to `channel_with_explicit_capacity`
that allows the caller to set the total amount of on-disk consumption.
The caller may choose to loop on a send which results in
`hopper::Error::Full` or drop the data and try again later. Hopper
will _not_ block sends when there is no space available -- in memory
or disk -- to store the item.

This work is done in the service of supporting
https://github.com/postmates/cernan/issues/411 and
https://github.com/postmates/cernan/issues/412.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>