Bump to 0.4.0
This release is hard-won. There should be significant performance
improvements for most workloads in terms of throughput and on-disk
consumption. End-users may now configure hopper to consume a fixed
amount of bytes in-memory and bytes on-disk, with appropriate error
returns to allow the end-user to recover.

This release has been tested on x86 and ARMv6.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>